REFACT: Tests: Moves class TestCase to top level, and renames test package
